+++ Application/Locators/LocatorStrategy.cs
@@ -1,3 +1,0 @@
-namespace Application.Locators;
-
-public sealed record LocatorStrategy();