+++ Application/Logging/IPrintStrategy.cs
@@ -1,0 +1,6 @@
+namespace Application.Logging;
+
+public interface IPrintStrategy
+{
+    void Render(LogEntry entry);
+}